Move index and bulk_update  actions to good ol' HTTP requests
We've found that we just can't rely in StimulusReflex (and underlying WebSockets) to guarantee a response to a request.
Because of this, there was intermittent issues when the server was overloaded with large requests, and the response never arrived, leaving an infinite loader, and a poor user wondering if anything was still happening.