Submit forms with Turbo Frame
Now the filters, pagination and product forms submit and load with Turbo.
Also the clear_search button doesn't load within the frame, no matter what attrs I give it. Not sure why.

And it's still building the whole page on the server.. I think we need Turbo Streams if we want to send back just a partial.