Submit product changes with Turbo Drive
This is not really faster, but a bit smoother because it avoids a full page render in the browser. The default Turbo loading indicator is shown (blue line at top).

We can't implement this for the filters or pagination because they use GET forms, which don't seem to be supported by Turbo Drive ((https://turbo.hotwired.dev/handbook/drive#redirecting-after-a-form-submission). Instead they will require Turbo Frames or Streams.

On to the next level!