Convert builder to instance method
Hmm, how does storing the graph in the OrderBuilder help?
It means I can create an #add_order_lines to automatically build and append order lines.

But I think the real benefit I'm after is _passing_ the graph object through to subsequent build_ calls. This could be done from the controller. So maybe rewind and do that first to see if it clarifies things.

(the new spec is helpful though; keep that)